Rename ets_context module to ets_event_loop
